% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_spec_to_tree_prior.R
\name{convert_spec_to_tree_prior}
\alias{convert_spec_to_tree_prior}
\title{Convert the \code{spec} to a tree prior}
\usage{
convert_spec_to_tree_prior(spec)
}
\arguments{
\item{spec}{the XML \code{spec} attribute}
}
\description{
Convert the \code{spec} to a tree prior
}
