% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_inference_model_from_file.R
\name{create_inference_model_from_file}
\alias{create_inference_model_from_file}
\title{Create an inference model from file}
\usage{
create_inference_model_from_file(xml_filename)
}
\arguments{
\item{xml_filename}{name of a BEAST2 XML input file,
as can be created by BEAUti2 or
using \link[beautier]{create_beast2_input_file_from_model}}
}
\description{
Create an inference model from file
}
