% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_inference_models_from_files.R
\name{create_inference_models_from_files}
\alias{create_inference_models_from_files}
\title{Create one or more BEAST2 inference models from one or more BEAST2
XML input files.}
\usage{
create_inference_models_from_files(xml_filenames)
}
\arguments{
\item{xml_filenames}{paths to one or more BEAST2 XML input files,
as can be created by BEAUti or beautier}
}
\description{
Create one or more BEAST2 inference models from one or more BEAST2
XML input files.
}
